requirements.txt
pandas
faiss-cpu
cohere
streamlit
torch
transformers
python-dotenv
tqdm